# [Winlogbeat] Add experimental event log reader with increased performance