--- a/winlogbeat/eventlog/wineventlog_expirimental.go
+++ b/winlogbeat/eventlog/wineventlog_expirimental.go
@@ -69,9 +69,7 @@ func (l *winEventLogExp) Open(state checkpoint.EventLogState) error {
 		win.WithSubscriptionFactory(func() (handle win.EvtHandle, err error) {
 			return l.open(l.lastRead)
 		}),
-		win.WithBatchSize(l.maxRead),
-		win.WithLogger(l.log))
-
+		win.WithBatchSize(l.maxRead))
 	return err
 }