@andrewkroh 2020-03-13
> Any reasons why you prefer not to enable it by default?

I want it to get some more testing before rolling it out to everyone by default. If everything goes well with it then we can make it the default in a future 7.x release. Then in 8.0 fully remove the old reader. This will provide an opportunity to remove a lot of the old code.