@SwiftOnSecurity 2020-12-14
Hello, what's the right way to report an issue with this experimental flag?
I can create a formal bug if you want.

I get these log entries in some custom Windows channels created by LogBinder SuperCharger 

WARN	[wineventlog.renderer]	wineventlog/renderer.go:319	The number of event data parameters doesn't match the number of parameters in the template.	{"channel": "Supercharger-Destination-WEC_ApplicationCrashes/Log", "provider": "Application Error", "event_id": 1000, "event_parameter_count": 13, "template_parameter_count": 15, "template_version": 0, "event_version": 0}

WARN	[wineventlog.renderer]	wineventlog/renderer.go:319	The number of event data parameters doesn't match the number of parameters in the template.	{"channel": "Supercharger-Destination-WEC_Firewall/Log", "provider": "Microsoft-Windows-Windows Firewall With Advanced Security", "event_id": 2004, "event_parameter_count": 23, "template_parameter_count": 26, "template_version": 0, "event_version": 0}

WARN	[wineventlog.renderer]	wineventlog/renderer.go:319	The number of event data parameters doesn't match the number of parameters in the template.	{"channel": "Supercharger-Destination-WEC_BITSClient/Log", "provider": "Microsoft-Windows-Bits-Client", "event_id": 59, "event_parameter_count": 9, "template_parameter_count": 10, "template_version": 1, "event_version": 1}

WARN	[wineventlog.renderer]	wineventlog/renderer.go:319	The number of event data parameters doesn't match the number of parameters in the template.	{"channel": "Supercharger-Destination-WEC_BITSClient/Log", "provider": "Microsoft-Windows-Bits-Client", "event_id": 60, "event_parameter_count": 15, "template_parameter_count": 17, "template_version": 1, "event_version": 1}