@andrewkroh 2020-12-15
Please open a new issue. See if you can grab the XML content of an event id 1000 sample from `Supercharger-Destination-WEC_ApplicationCrashes/Log` using Windows Event Viewer. Also take a look at the metadata that the provider registered. You can use `wevutil get-publisher` to find the definition for event ID 1000 IIRC.